Adds Hamilton to README.md
I think this would be a nice fit for the orchestration section. Happy to tune the wording if you want. Its different from other orchestrators as its lightweight and more fine-grained -- encourages the user to think in terms of the logical shape of the pipeline rather than the infrastructure required to run it.